requirements.txt
# requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
notion-client>=2.0.0
google-generativeai>=0.8.0
openai>=1.12.0
python-dotenv>=1.0.0
tqdm>=4.66.0
